--- src/utils/useModal.js
@@ -1,13 +1,0 @@
-import { useState } from 'react';
-const useModal = () => {
-  const [isVisible, setIsVisible] = useState(false);
-
-  function toggleModal() {
-    setIsVisible(!isVisible);
-  }
-  return {
-    isVisible,
-    toggleModal,
-  };
-};
-export default useModal;
